core/auth/viewsets/refresh.py
from rest_framework.response import Response
from rest_framework import viewsets
from rest_framework.permissions import AllowAny
from rest_framework import status
from rest_framework_simplejwt.exceptions import TokenError, InvalidToken
from rest_framework_simplejwt.views import TokenRefreshView


class RefreshViewSet(viewsets.ViewSet, TokenRefreshView):
    permission_classes = (AllowAny, )
    http_method_names = ['post']

    def create(self, request, *args, **kwargs):
        serializer = self.get_serializer(data=request.data)

        try:
            serializer.is_valid(raise_exception=True)
        except TokenError as e:
            raise InvalidToken(e.args[0])

        return Response(serializer.validated_data, status=status.HTTP_200_OK)

    def http_method_not_allowed(self, request, *args, **kwargs):
        allowed_methods = ', '.join(self.http_method_names)
        return Response(
            {
                "detail": f"Only {allowed_methods} method(s) are/is allowed."
            },
            status.HTTP_405_METHOD_NOT_ALLOWED
        )