requirements/requirements.txt
alabaster==0.7.16
asgiref==3.7.2
astroid==3.0.3
Babel==2.14.0
build==1.0.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.8
Django==4.0.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docutils==0.20.1
drf-nested-routers==0.93.5
drf-yasg==1.21.7
flake8==6.1.0
flake8-django==1.4
idna==3.6
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.3
lazy-object-proxy==1.10.0
MarkupSafe==2.1.4
mccabe==0.7.0
openapi-codec==1.3.2
packaging==23.2
pillow==10.2.0
pip-tools==7.3.0
platformdirs==4.2.0
pluggy==1.3.0
psycopg2-binary==2.9.3
pycodestyle==2.11.1
pydotplus==2.0.2
pyflakes==3.1.0
Pygments==2.17.2
PyJWT==2.8.0
pylint==3.0.3
pyparsing==3.1.1
pyproject_hooks==1.0.0
pytest==7.4.4
pytest-django==4.8.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.4.4
tomlkit==0.12.3
uritemplate==4.1.1
urllib3==2.1.0
wrapt==1.16.0
